+++ lib/extended_email_reply_parser/mail/message.rb
@@ -20,5 +20,9 @@ module Mail
     def extract_text
       (self.text_part || (self if self.content_type.include?('text/plain'))).body_in_utf8
     end
+
+    def parse
+      ExtendedEmailReplyParser.parse self
+    end
   end
 end